requirements.txt
django==4.2.7
psycopg2-binary==2.9.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-auth-ldap==4.6.0
python-ldap==3.4.3
django-extensions==3.2.3
pillow==10.0.1
python-decouple==3.8
